% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_uni_outliers.R
\name{acc_uni_outliers}
\alias{acc_uni_outliers}
\title{Assess Accuracy - Univaritate Outliers}
\usage{
acc_uni_outliers(data, metadata)
}
\arguments{
\item{data}{\itemize{
\item arrow data table
}}

\item{metadata}{\itemize{
\item item level metadata, expected data table object generated by prep_metadata function
}}
}
\value{
list of 2 items:
result - result summary data.frame
outliers - return outliers values for each variable
}
\description{
check for univariate outliers
definition of outliers: based on method introduced by Tukey in 1977
\itemize{
\item less than 1.5*IQR away from 1st quantile
\item greater than 1.5*IQR away from the 3rd quantile
}
}
