% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_arrow_summarize.R
\name{arrow_summary}
\alias{arrow_summary}
\title{Custom summary function for arrow object}
\usage{
arrow_summary(data, var, ...)
}
\arguments{
\item{data}{\itemize{
\item arrow data table to analyze, can be passed via pipe
}}

\item{var}{\itemize{
\item variables to generate summary
}}

\item{...}{}
}
\description{
Custom summary function for arrow object
}
