% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_crude_missing.R
\name{com_crude_missing}
\alias{com_crude_missing}
\title{Assess Completeness - Crude Missing}
\usage{
com_crude_missing(data, item_metadata, cross_item_metadata)
}
\arguments{
\item{data}{\itemize{
\item arrow data table
}}

\item{cross_item_metadata}{\itemize{
\item cross level metadata,  expected data table object generated by prep_metadata function
}}

\item{metadata}{\itemize{
\item item level metadata, expected data table object generated by prep_metadata function
}}
}
\value{
2 data.frame for univariate and multivariate missingness result
}
\description{
Assess Completeness - Crude Missing
}
