% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_qualified_missing.R
\name{com_qualified_missing}
\alias{com_qualified_missing}
\title{Assess Completeness - Qualified missing}
\usage{
com_qualified_missing(data, metadata)
}
\arguments{
\item{data}{\itemize{
\item arrow data table
}}

\item{metadata}{\itemize{
\item item level metadata, expected data table object generated by prep_metadata function
}}
}
\value{
result summary data.frame
}
\description{
Assess Completeness - Qualified missing
}
