% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_contradiction.R
\name{con_contradiction}
\alias{con_contradiction}
\title{Assess Consistency - Value Contradiction
Summarize how much data has contradicting values, and return the rows with contradicting values}
\usage{
con_contradiction(data, metadata)
}
\arguments{
\item{data}{\itemize{
\item arrow data table
}}

\item{metadata}{\itemize{
\item item level metadata, expected data table object generated by prep_metadata function
}}
}
\value{
list containing result summary (data.frame) and contradicted data (tibble)
}
\description{
Assess Consistency - Value Contradiction
Summarize how much data has contradicting values, and return the rows with contradicting values
}
