% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_label.R
\name{con_label}
\alias{con_label}
\title{Assess Consistency - Valid labels for categorical labels}
\usage{
con_label(data, metadata, path = NULL)
}
\arguments{
\item{data}{\itemize{
\item arrow data table
}}

\item{metadata}{\itemize{
\item item level metadata, expected data table object generated by prep_metadata function
}}

\item{path}{\itemize{
\item path to item level metadata
}}
}
\value{
result summary for consistency assessment
}
\description{
Assess Consistency - Valid labels for categorical labels
}
