% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_range.R
\name{con_range}
\alias{con_range}
\title{Assess Consistency - Range of values}
\usage{
con_range(data, metadata)
}
\arguments{
\item{data}{\itemize{
\item arrow data table
}}

\item{metadata}{\itemize{
\item item level metadata, expected data table object generated by prep_metadata function
}}
}
\value{
result summary data.frame
}
\description{
Assess Consistency - Range of values
}
