% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_datatype.R
\name{int_datatype}
\alias{int_datatype}
\title{Assess Integrity - Datatype \n}
\usage{
int_datatype(data, metadata)
}
\arguments{
\item{data}{\itemize{
\item arrow data table
}}

\item{metadata}{\itemize{
\item item level metadata, expected data table object generated by prep_metadata function
}}
}
\description{
Check whether all variables follow the pre-defined datatype,
and cast the given data to the defined datatype
}
