% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_duplicates.R
\name{int_duplicates}
\alias{int_duplicates}
\title{Assess Integrity - duplication
Check whether there are duplicate rows in pre-defined columns/variables}
\usage{
int_duplicates(
  data,
  check_all = TRUE,
  remove_dups = FALSE,
  cross_item_metadata = NULL
)
}
\arguments{
\item{data}{\itemize{
\item arrow data table
}}

\item{check_all}{\itemize{
\item whether to check duplicates over all variables
}}

\item{remove_dups}{\itemize{
\item whether to remove duplicates
}}

\item{cross_item_metadata}{\itemize{
\item cross item level metadata, expected data table object generated by prep_metadata function
}}
}
\value{
result data.frame
}
\description{
Assess Integrity - duplication
Check whether there are duplicate rows in pre-defined columns/variables
}
