% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinsic_eval.R
\name{intrinsic_eval}
\alias{intrinsic_eval}
\title{Perform all Intrinsic quality assessment}
\usage{
intrinsic_eval(metadata_path, data)
}
\arguments{
\item{metadata_path}{\itemize{
\item path to the data folder
}}

\item{data}{\itemize{
\item arrow data table
}}
}
\description{
Perform all Intrinsic quality assessment
}
