% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_mapping_filter_condition.R
\name{mapping_filter_condition}
\alias{mapping_filter_condition}
\title{Mapping value range operator to the built in function}
\usage{
mapping_filter_condition(
  data,
  var,
  datatype,
  lower_bound,
  higher_bound,
  greater,
  less
)
}
\arguments{
\item{data}{\itemize{
\item arrow data table
}}

\item{var}{\itemize{
\item variable for fitlering
}}

\item{datatype}{\itemize{
\item datatype of variable
}}

\item{lower_bound}{\itemize{
\item lowerbound value
}}

\item{higher_bound}{\itemize{
\item higherbound value
}}

\item{greater}{\itemize{
\item "(" or "["
}}

\item{less}{\itemize{
\item ")" or "]"
}}
}
\description{
Mapping value range operator to the built in function
}
