% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_metadata.R
\name{prep_crossitem_metadata}
\alias{prep_crossitem_metadata}
\title{Prepare cross-item level metadata}
\usage{
prep_crossitem_metadata(path = NULL)
}
\arguments{
\item{path}{\itemize{
\item path to the metadata folder
}}
}
\value{
a list of 2 data.table:
\itemize{
\item multivariates - define multivariate variables
\item contradiction - contradiction rules
}
}
\description{
Read cross-item level metadata
}
