% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_metadata.R
\name{prep_item_metadata}
\alias{prep_item_metadata}
\title{Prepare item level metadata}
\usage{
prep_item_metadata(path = NULL)
}
\arguments{
\item{path}{\itemize{
\item path to the metadata folder
}}
}
\value{
a data.table for item level metadata
}
\description{
Read item level metadata
}
