% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_compute_duplicates.R
\name{util_compute_duplicates}
\alias{util_compute_duplicates}
\title{Utility to check each row whether they are duplicate or not}
\usage{
util_compute_duplicates(data, vars)
}
\arguments{
\item{data}{\itemize{
\item arrow data table
}}

\item{vars}{\itemize{
\item variables to check for duplicates
}}
}
\value{
logical vector indicating whether each row is a duplicate or not
}
\description{
Utility to check each row whether they are duplicate or not
}
