% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_compute_outliers.R
\name{util_compute_outliers}
\alias{util_compute_outliers}
\title{Compute outliers}
\usage{
util_compute_outliers(data, var)
}
\arguments{
\item{data}{}

\item{var}{\itemize{
\item variable to calculate outlier, expected to be a numeric variable
}}
}
\value{
arrow table containing outliers
}
\description{
Compute outliers
}
