% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_export.R
\name{util_export}
\alias{util_export}
\title{Utility function - Export data}
\usage{
util_export(data, outfile_name = "failed_check_data")
}
\arguments{
\item{data}{\itemize{
\item data returned by evaluation functionss
}}

\item{outfile_name}{\itemize{
\item name of the output file
}}
}
\description{
Save data that failed the checks as an excel file
}
