% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_graphing.R
\name{util_graphing_percentage}
\alias{util_graphing_percentage}
\title{Title}
\usage{
util_graphing_percentage(data, label, percentage)
}
\arguments{
\item{data}{\itemize{
\item the result dataframe generated by DQA functions
}}

\item{label}{\itemize{
\item the column used for labeling (usually varname column)
}}

\item{percentage}{\itemize{
\item column containing percentage to graph
}}
}
\description{
Title
}
