% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_is_integer.R
\name{util_is_integer}
\alias{util_is_integer}
\title{Utility to check whether a vector only contains integer}
\usage{
util_is_integer(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{\itemize{
\item object to †est
}}

\item{tol}{\itemize{
\item precision of the detection. Values deviating more than \code{tol} from
their closest integer value will not be deemed integer.
}}
}
\value{
boolean value of whether vector only contains integer
}
\description{
Utility to check whether a vector only contains integer
}
