% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_mapping_compare.R
\name{util_mapping_compare}
\alias{util_mapping_compare}
\title{Utility to map operator to comparison function}
\usage{
util_mapping_compare(data, var1, var2, operator)
}
\arguments{
\item{data}{}

\item{var1}{}

\item{var2}{}

\item{operator}{}
}
\description{
Utility to map operator to comparison function
}
